# Refactor LazySegmentTree to be easier to modify