@Mukundan314 2024-11-06
Example usage: https://codeforces.com/contest/2027/submission/290258327 (lazy segment tree is overkill for this problem but demonstrates the usage)